requirements.txt
PyQt6
google-generativeai
pandas